+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 
 # 0.1.2
  - Adding to_i and to_f as preprocessors
+
+# 0.1.3
+ - Adding strip as preprocessor
+ - Allowing multiple preprocessors